Apply suggestions from code review

+++ modin/core/dataframe/pandas/dataframe/dataframe.py
@@ -991,17 +991,17 @@ class PandasDataframe(ClassLogger):
             return self.copy()
 
         if col_positions is None and row_positions is not None:
-            # This way tries to eliminate the optimization that first takes part of the data using
-            # the mask operation, so that later less data is concated into a whole column.
-            # Note: in the case where a small part of the data is discarded, the overhead of the
-            # engines (for putting data in and out of storage) can exceed the resulting speedup.
+            # Check if the optimization that first takes part of the data using the mask 
+            # operation so that later less data is concatenated into a whole column is useful.
+            # In the case when only a small portion of the data is discarded, the overhead of the
+            # engine (for putting data in and out of storage) can exceed the resulting speedup.
             all_rows = None
             if self.has_materialized_index:
                 all_rows = len(self.index)
             elif self._row_lengths_cache:
                 all_rows = sum(self._row_lengths_cache)
             if all_rows:
-                if len(row_positions) > int(0.9 * all_rows):
+                if len(row_positions) > 0.9 * all_rows:
                     return self._reorder_labels(
                         row_positions=row_positions, col_positions=col_positions
                     )
